pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trip-agent"
version = "1.0.0"
description = "AI 旅行行程规划助手 — LangGraph + Pydantic + 通义千问"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "openai>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "requests>=2.31.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "httpx>=0.27.0",
]
retrieval = [
    "faiss-cpu",
    "sentence-transformers",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
